Move the "convert path to absolute" logic from PathArc to PathAbs.
